requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9
sqlalchemy==2.0.27
pydantic==2.6.3
pydantic-settings==2.2.1
easyocr==1.7.1
pymupdf==1.23.26
opencv-python-headless==4.9.0.80
sentence-transformers==2.5.1
faiss-cpu==1.8.0
groq>=0.5.0
python-dotenv==1.0.1
numpy<2
